Update to buttress/config
Changed the way values were stored. Previously, there was a map
per type, which made adding a new type a little cumbersome. Now
there is a single map with a compound key that includes type.

Changed the getters to return an error if the value is undefined
and added a MustGet method for each that panics if the value is
undefined.

Added bool, float and int types.